requirements.txt
pytesseract
pillow
nltk
numpy